style(daemon): rename socket-path.ts -> paths.ts to match its Paths namespace
The module is imported as `import * as Paths`, so the filename should be
paths.ts (filename<->identifier law). Test renamed alongside.
