+++ lib/commands/daemon/paths.ts
